\name{DatabaseConnect}
\alias{DatabaseConnect}
\title{Connect to a database and view it...}
\usage{DatabaseConnect(toolkit="RGtk2")}
\description{Connect to a database and view it}
\details{Provides a connection dialog and a viewer to explore a MySQL database}
\author{Dason Kurkiewicz \email{dasonk@iastate.edu}}
\arguments{\item{toolkit}{Which gWidgets toolkit to use when creating the GUI}}
