\name{sqlToR}
\alias{sqlToR}
\title{Convert data types from SQL to R...}
\usage{sqlToR(data)}
\description{Convert data types from SQL to R}
\details{Takes a vector of SQL data types and and returns the corresponding
data types in R.}
\value{A vector of R data types.}
\author{Dason Kurkiewicz \email{dasonk@iastate.edu}}
\arguments{\item{data}{The vector of SQL data types to convert}}
