\name{unix2POSIXct}
\alias{unix2POSIXct}
\title{Convert times to a more useful form...}
\usage{unix2POSIXct(time, type="second")}
\description{Convert times to a more useful form}
\details{Converts times to a more useful form}
\value{A more useful time}
\arguments{\item{time}{A single time}
\item{type}{Either 'second' or 'millisecond' specifying whether the data is seconds
or milliseconds from the epoch}}
